Create separate component for JoblistingItem
To avoid using the component from a different route, we rather create a
separate component for it. Then this one is used in CompanyDetail and in
JoblistingList.
