+++ app/routes/joblistings/components/JoblistingList.css
@@ -28,62 +28,3 @@
     margin-right: 10px;
   }
 }
-
-.joblistingItem {
-  display: flex;
-  align-items: center;
-  margin-top: 3px;
-  margin-bottom: 12px;
-  padding: 15px;
-  line-height: 1.3;
-  min-width: 280px;
-  border-radius: 5px;
-  color: var(--lego-font-color);
-  transition: background-color var(--easing-fast);
-
-  &:hover {
-    background-color: rgba(255, 0, 0, var(--color-red-hover-alpha));
-  }
-
-  @media (--mobile-device) {
-    padding: 5px;
-  }
-}
-
-.companyLogo {
-  width: 120px;
-  max-height: 80px;
-  min-width: 120px;
-  object-fit: contain;
-  margin-right: 20px;
-  background: white;
-}
-
-.listItem {
-  display: flex;
-  flex: 1;
-  justify-content: space-between;
-  font-size: 15px;
-
-  @media (--mobile-device) {
-    flex-direction: column;
-  }
-}
-
-.joblistingItemTitle {
-  font-weight: 600;
-  font-size: 16px;
-  margin-right: 5px;
-  margin-bottom: 10px;
-}
-
-.deadLine {
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-  min-width: 145px;
-
-  @media (--mobile-device) {
-    justify-content: flex-start;
-  }
-}
