+++ app/routes/joblistings/components/JoblistingRightNav.tsx
@@ -4,8 +4,8 @@ import { useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
 import Button from 'app/components/Button';
 import { CheckBox, RadioButton } from 'app/components/Form/';
+import type { JobType } from 'app/components/JoblistingItem/Items';
 import type { ActionGrant } from 'app/models';
-import type { JobType } from 'app/routes/joblistings/components/Items';
 import { jobTypes } from '../constants';
 import styles from './JoblistingRightNav.css';
